+++ codex-rs/core/src/state/mod.rs
@@ -7,4 +7,3 @@ pub(crate) use session::SessionState;
 pub(crate) use turn::ActiveTurn;
 pub(crate) use turn::RunningTask;
 pub(crate) use turn::TaskKind;
-pub(crate) use turn::TurnTimingState;
